\name{getValueFromScript}
\alias{getValueFromScript}
\title{Get Method}
\usage{
getValueFromScript(x, name, coercion = TRUE, ...)
}
\arguments{
  \item{x}{a S3 \code{ParsedGroovyScriptIOVar} object}

  \item{name}{name of the variablelogical value.}

  \item{coercion}{logical value. If it is \code{TRUE}
  (Deafault) the returened variable is coerced.}

  \item{...}{further arguments}
}
\description{
Gets the value of a variable in a S3
\code{ParsedGroovyScriptIOVar} object
}

